requirements.txt
flask==2.0.1
tweepy==4.14.0
python-dotenv==0.19.0
requests==2.26.0
